+++ libtac/lib/connect.c
@@ -32,7 +32,7 @@
 #include "libtac.h"
 
 /* Pointer to TACACS+ connection timeout */
-int tac_timeout = 5;
+unsigned long tac_timeout = 5;
 
 /* Returns file descriptor of open connection
    to the first available server from list passed
